Devlp3.2 (#8)
* Программа на Java, реализация нормализованного кол-ва элементов, вариант 3

* Laba2
